--- a/luarules/gadgets/game_commander_builder.lua
+++ b/luarules/gadgets/game_commander_builder.lua
@@ -45,8 +45,8 @@ function SpawnAssistTurret(unitID, unitDefID, unitTeam)
     local spawnpadunit = spawnpads[unitDefID]
     local spawnpadID
     for k = 1,10000 do
-        posx = posx + math.random(-k-64, k+64)
-        posz = posz + math.random(-k-64, k+64)
+        posx = math.ceil((posx + math.random(-k-64, k+64))/16)*16
+        posz = math.ceil((posz + math.random(-k-64, k+64))/16)*16
         posy = Spring.GetGroundHeight(posx, posz)
         local canSpawnTurret = positionCheckLibrary.FlatAreaCheck(posx, posy, posz, 96)
         if canSpawnTurret then