--- a/luarules/gadgets/unit_juno_damage.lua
+++ b/luarules/gadgets/unit_juno_damage.lua
@@ -48,6 +48,8 @@ if gadgetHandler:IsSyncedCode() then
 		['corspec'] = true,
 		['corvoyr'] = true,
 		['corvrad'] = true,
+		['legarad'] = true,
+		['legajam'] = true,
 		
 		['armmine1'] = true,
 		['armmine2'] = true,