--- a/scripts/Units/armspy_clean.bos
+++ b/scripts/Units/armspy_clean.bos
@@ -277,7 +277,25 @@ HitByWeapon(anglex, anglez)
 	return (0);
 }
 
+AimWeapon1()
+{
+	return (TRUE);
+}
 
+FireWeapon1()
+{
+	get KILL_UNIT(get MY_ID,TRUE,FALSE);
+}
+
+QueryWeapon1(piecenum)
+{
+	piecenum = body;
+}
+
+AimFromWeapon1(piecenum)
+{
+	piecenum = body;
+}
 
 Killed(severity, corpsetype)
 {