--- a/scripts/Units/corban_clean.bos
+++ b/scripts/Units/corban_clean.bos
@@ -96,8 +96,6 @@ AimWeapon1(heading, pitch)
 	// logic to reset gun to neutral if the target heading has been
 	// out of firing arc for ~3 seconds
 	ourheading = heading;
-	if (heading < -24576) ourheading = -24576;
-	if (heading > 24576) ourheading = 24576;
 
 
 	turn turret to y-axis ourheading speed <125.0>;