--- a/units/ArmBots/T2/armsptk.lua
+++ b/units/ArmBots/T2/armsptk.lua
@@ -146,7 +146,7 @@ return {
 					cruise_min_height = "15",
 					lockon_dist = "150",
 					overrange_distance = 690,
-￼					projectile_destruction_method = "descend",
+					projectile_destruction_method = "descend",
 					speceffect = "cruise",
 					when = "distance>0",
 				},