--- a/units/ArmBuildings/LandEconomy/armsolar.lua
+++ b/units/ArmBuildings/LandEconomy/armsolar.lua
@@ -3,7 +3,7 @@ return {
 		activatewhenbuilt = true,
 		buildangle = 4000,
 		buildpic = "ARMSOLAR.DDS",
-		buildtime = 2800,
+		buildtime = 2600,
 		canrepeat = false,
 		collisionvolumeoffsets = "0.0 -18.0 1.0",
 		collisionvolumescales = "50.0 76.0 50.0",
@@ -12,7 +12,6 @@ return {
 		damagemodifier = 0.5,
 		energycost = 0,
 		energystorage = 50,
-		-- negative energyupkeep means it's conditional so when unit is off won't produce
 		energyupkeep = -20,
 		explodeas = "smallBuildingexplosiongeneric",
 		footprintx = 5,