--- a/units/CorAircraft/corca.lua
+++ b/units/CorAircraft/corca.lua
@@ -34,7 +34,7 @@ return {
 		terraformspeed = 225,
 		turninplaceanglelimit = 360,
 		turnrate = 240,
-		workertime = 60,
+		workertime = 65,
 		buildoptions = {
 			[1] = "corsolar",
 			[2] = "coradvsol",