--- a/units/CorBots/T2/corack.lua
+++ b/units/CorBots/T2/corack.lua
@@ -37,7 +37,7 @@ return {
 		turninplacespeedlimit = 0.726,
 		turnrate = 1075.25,
 		upright = true,
-		workertime = 180,
+		workertime = 190,
 		buildoptions = {
 			[1] = "corfus",
 			[2] = "corafus",