--- a/units/CorBuildings/LandDefenceOffence/corllt.lua
+++ b/units/CorBuildings/LandDefenceOffence/corllt.lua
@@ -132,15 +132,15 @@ return {
 				turret = true,
 				weapontype = "BeamLaser",
 				weaponvelocity = 2250,
+				customparams = {
+					exclude_preaim = true,
+				},
 				damage = {
 					commanders = 112.5,
 					default = 75,
 					subs = 5,
 					vtol = 5,
 				},
-				customparams = {
-					exclude_preaim = true
-				}
 			},
 		},
 		weapons = {