--- a/units/CorVehicles/T2/cortrem.lua
+++ b/units/CorVehicles/T2/cortrem.lua
@@ -112,7 +112,7 @@ return {
 		weapondefs = {
 			tremor_spread_fire = {
 				accuracy = 0,
-				areaofeffect = 200,
+				areaofeffect = 210,
 				avoidfeature = false,
 				cegtag = "arty-medium",
 				craterboost = 0,
@@ -134,13 +134,13 @@ return {
 				weapontype = "Cannon",
 				weaponvelocity = 420,
 				customparams = {
-					max_range_reduction = "0.28",
+					max_range_reduction = "0.30",
 					speceffect = "sector_fire",
-					spread_angle = "20",
+					spread_angle = "22",
 					when = "always",
 				},
 				damage = {
-					default = 150,
+					default = 200,
 					subs = 75,
 					vtol = 19,
 				},