--- a/weapons/spybombx.lua
+++ b/weapons/spybombx.lua
@@ -20,7 +20,7 @@ return {
 		tolerance = 7000,
 		turret = 1,
 		damage = {
-			default = 4500,
+			default = 5000,
 		},
 	},
 	spybombxscav = {