… scripts/Units/armfido_clean.cob
… scripts/Units/armkam.cob
… scripts/Units/armkam_clean.cob
… scripts/Units/armspy_clean.cob
… scripts/Units/corban.cob
… scripts/Units/corban_clean.cob
… scripts/Units/corspy_clean.cob
… unitbasedefs/proposed_unit_reworks_defs.lua
… units/ArmAircraft/armkam.lua
… units/ArmBots/T2/armspy.lua
… units/ArmBots/armpw.lua
… units/ArmBuildings/LandUtil/armarad.lua
… units/ArmBuildings/LandUtil/armjamt.lua
… units/ArmShips/T2/armsjam.lua
… units/ArmVehicles/armflash.lua
… units/CorBots/T2/corspy.lua
… units/CorBots/corak.lua
… units/CorBuildings/LandEconomy/corwin.lua
… units/CorBuildings/LandUtil/corarad.lua
… units/CorVehicles/T2/corparrow.lua