Changelist:

General
  Laser weapons no longer randomly miss vs moving targets

Economy
  Arm solar 2800 -> 2600 buildtime
  Arm wind 37 -> 40m
  Cortex wind 45 -> 43m, 199 -> 220hp
  Arm tidal 250 -> 200e
  Arm asolar 370 -> 350m
  Cortex cons +5% bp, since they cost more
  Exploiter buildtime +7%

T2 radars
  Metalcost 560 -> 400
  Energycost 19000 -> 14000
  Buildtime 11800 -> 8000

Sneaky Pete
  Cloaking removed

Jammer ships
  Costs +130%
  Health +130%
  Speed 61 -> 40


Spybots
  Cloaking cost 100/50 -> 40/15
  EMP immunity removed
  Attack command for self-destruct added
  Damage 56000 -> 5000
  Paralyzetime 35/20 -> 8s

Banisher
  Can now fire in all directions
  Sightdistance 550 -> 650
  Energycost 23000 -> 17000
  Improved aim/tracking

Tremor
  Damage 150 -> 200
  AoE 200 -> 210
  Weapon hits a wider area

Poison Arrow
  Energycost 29000 -> 21000
  Buildtime 22200 -> 19000
  Sightdistance 385 -> 500

Turtle
  Sightdistance 372 -> 500

Garpike, Pincer
  Sightdistance 305 -> 500
  Faster turning turrets

Hound
  Lost gauss weapon
  Metalcost 300 -> 285

Recluse
  Range 600 -> 575
  Speed 52 -> 47


Grunt
  Metalcost 36 --> 43
  Energycost 880 -> 840
  Range 230 -> 210
  Turnrate 1391 -> 1200

Pawn
  Metalcost 48 -> 52
  Energycost 960 -> 870
  Health 335 -> 370

Tick
  Metalcost 17 -> 21
  Energycost 340 -> 300
  Health 61 -> 60

Scout cars
  Health +20%

Blitz
  Health 690 -> 750
  Cost rounded: 109m, 911e -> 110m, 900e
  Ellipsoid hitbox
  Sightrange 299 -> 350

Incisor
  Cost rounded: 118m, 1042e -> 120m, 1040e
  Buildtime 1761 -> 2200
  Sightrange 273 -> 330

Shuriken
  Reloadtime 1.2s -> 1.3s

Stormbringer
  Drops bombs 17% closer together

Banshee
  Health 485 -> 560
  Altered weapon, plasma bullets with similar stats
  Turnrate increased