@drivver44 2025-03-08
You miss understood my point about historical hitboxes
The issues of the historical hitboxes was that I could not find reasons why
they were defined as they were.
Atleast not on GitHub
We now have a document that controls how hitboxes can look
As for the seperate pr
The issue is that you are adding a whole slew of balance changes with a
technical change.
If the pr has an issue where it needs to get reverted
We lose the hitbox as defined here and revert back to the hitbox you do not
like.
That’s why I’m asking for a seperate or on the hitbox change alone.
Idm if you wish to add more hitboxes in that or if you wish to add more.

As for the use of Elli please review the internal document and find out how
to better define when Eli can and cannot be used.
I’m pretty sure we are on the same page.


On Fri, Mar 7, 2025 at 1:03 PM Mitvit ***@***.***> wrote:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In units/ArmVehicles/armflash.lua
> <https://github.com/beyond-all-reason/Beyond-All-Reason/pull/4450#discussion_r1985708684>
> :
>
> > @@ -3,15 +3,15 @@ return {
>  		buildpic = "ARMFLASH.DDS",
>  		buildtime = 1960,
>  		canmove = true,
> -		collisionvolumeoffsets = "0 -2 -1",
> -		collisionvolumescales = "24 13 31",
> -		collisionvolumetype = "Box",
> +		collisionvolumeoffsets = "0 -3 0",
>
> Historical hitboxes don't matter, we should focus on what we have now.
>
> The basic design goal is reasonably clear, as told by Seth above.
>
> When I said different hitbox schemes, I mean that it's fine for some units
> to have a box and some have an ellipsoid - whatever gets the job done. Not
> that some units should randomly have hitbox not covering their turret at
> all, or be off-size for their model.
>
> Right now the issue is that Blitz does a lot of friendly fire, with its
> 2-box hitbox. This is mitigated a ton by a change to an ellipsoid hitbox.
> This would likely be good also for the missile trucks, and then tigers,
> etc, or potentially most vehicles - but Blitz is the most urgent one to
> improve.
>


>